# spring-boot-example/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Pet Store API
  description: A sample API for managing pets
  version: 1.0.0
  contact:
    name: API Support
    email: api@example.com
    url: https://example.com/support
  
servers:
  - url: https://api.example.com/v1/petstore
    description: Production server
    
tags:
  - name: pets
    description: Pet operations
  - name: stores
    description: Store operations
    
paths:
  /pets:
    get:
      description: Returns a list of all available pets
      tags:
        - pets
      operationId: getAllPets
      responses:
        '200':
          description: A list of pets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
              examples:
                petsExample:
                  value:
                    - id: 1
                      name: Fluffy
                      status: available
                    - id: 2
                      name: Rex
                      status: sold
  
  /pets/{petId}:
    get:
      description: Returns a pet by ID
      tags:
        - pets
      operationId: getPetById
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: A pet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
              examples:
                petExample:
                  value:
                    id: 1
                    name: Fluffy
                    status: available
                    
components:
  schemas:
    Pet:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        status:
          type: string
          enum:
            - available
            - pending
            - sold
      required:
        - id
        - name 